+++ src/server/server.py
@@ -29,7 +29,7 @@ import lib.tools as tools
 # DEBUG
 # from pdb import set_trace as st
 
-VERSION = '1.12.2'
+VERSION = '2.0.0'
 
 SERVER_OPTS, ARGS, TOOLS = tools.loadconfig(version=VERSION)
 
@@ -164,20 +164,6 @@ class Admin():
                 pg_conn.close()
                 return tools.response_render(
                     'OK: %s=%s for %s' % (key, value, username))
-            # Deprecated endpoint for principals
-            if key == 'principals':
-                value = unquote_plus(value)
-                cur.execute(
-                    """
-                    UPDATE USERS SET PRINCIPALS=(%s) WHERE NAME=(%s)
-                    """, (value, username))
-                pg_conn.commit()
-                cur.close()
-                pg_conn.close()
-                return tools.response_render(
-                    'WARNING: ' + \
-                    'This endpoint is deprecated, upgrade your client cassh to 1.7.0\n'+
-                    'OK: %s=%s for %s' % (key, value, username))
         return tools.response_render('WARNING: No key found...')
 
     def DELETE(self, username):
@@ -270,7 +256,8 @@ class Client():
         force_sign = False
 
         # LDAP ADMIN authentication
-        is_admin_auth, _ = tools.ldap_authentification(SERVER_OPTS, admin=True)
+        is_admin_auth, message = tools.ldap_authentification(
+            SERVER_OPTS, admin=True)
 
         payload, message = tools.data2map()
         if message:
@@ -355,7 +342,11 @@ class Client():
 
         status = user[2]
         expiry = user[3]
-        principals = tools.get_principals(user[4], username, shell=True)
+        custom_principals = tools.clean_principals_output(user[4], username, shell=True)
+        list_membership, _ = tools.get_memberof(
+            realname,
+            SERVER_OPTS)
+        full_principals = tools.merge_principals(custom_principals, list_membership, SERVER_OPTS)
 
         if status > 0:
             cur.close()
@@ -364,7 +355,7 @@ class Client():
             return tools.response_render("Status: %s" % constants.STATES[user[2]])
 
         cert_contents = TOOLS.sign_key(
-            tmp_pubkey.name, username, expiry, principals, db_cursor=cur)
+            tmp_pubkey.name, username, expiry, full_principals, db_cursor=cur)
 
         remove(tmp_pubkey.name)
         pg_conn.commit()
@@ -548,35 +539,6 @@ class Principals():
     """
     Class Principals
     """
-    def GET(self, username):
-        """
-        Get a user principals
-        """
-        # LDAP authentication
-        is_admin_auth, message = tools.ldap_authentification(SERVER_OPTS, admin=True)
-        if not is_admin_auth:
-            return tools.response_render(message, http_code='401 Unauthorized')
-
-        pg_conn, message = TOOLS.pg_connection()
-        if pg_conn is None:
-            return tools.response_render(message, http_code='503 Service Unavailable')
-        cur = pg_conn.cursor()
-        values = {'username': username}
-        cur.execute(
-            """
-            SELECT PRINCIPALS FROM USERS WHERE NAME=(%(username)s)
-            """, values)
-        principals = cur.fetchone()
-        pg_conn.commit()
-        cur.close()
-        pg_conn.close()
-        if not principals:
-            return tools.response_render(
-                "ERROR: {} doesn't exist or doesn't have principals...".format(
-                    username),
-                http_code='400 Bad Request')
-        return tools.response_render('OK: {} principals are {}'.format(username, principals))
-
     def POST(self, username):
         """
         Manage user principals
@@ -607,7 +569,7 @@ class Principals():
         values = {'username': username}
         cur.execute(
             """
-            SELECT NAME,PRINCIPALS FROM USERS WHERE NAME=(%(username)s)
+            SELECT NAME,PRINCIPALS,REALNAME FROM USERS WHERE NAME=(%(username)s)
             """, values)
         user = cur.fetchone()
         # If user dont exist
@@ -654,8 +616,14 @@ class Principals():
             elif key == 'purge':
                 values['principals'] = username
 
-        # Remove duplicates
-        values['principals'] = ','.join(list(dict.fromkeys(values['principals'].split(','))))
+        list_membership, _ = tools.get_memberof(
+            user[2],
+            SERVER_OPTS)
+        values['principals'] = tools.truncate_principals(
+            values['principals'],
+            list_membership,
+            SERVER_OPTS)
+
         cur.execute(
             """
             UPDATE USERS SET PRINCIPALS=(%(principals)s) WHERE NAME=(%(username)s)
@@ -663,6 +631,13 @@ class Principals():
         pg_conn.commit()
         cur.close()
         pg_conn.close()
+
+        # Add LDAP principals
+        values['principals'] = tools.merge_principals(
+            values['principals'],
+            list_membership,
+            SERVER_OPTS)
+
         return tools.response_render(
             "OK: {} principals are '{}'".format(username, values['principals']))
 
@@ -696,25 +671,49 @@ class PrincipalsSearch():
 
         cur.execute(
             """
-            SELECT NAME,PRINCIPALS FROM USERS
+            SELECT NAME,PRINCIPALS,REALNAME FROM USERS
             """)
         all_principals = cur.fetchall()
         pg_conn.commit()
         cur.close()
         pg_conn.close()
 
+        if SERVER_OPTS['ldap']:
+            ldap_conn, _ = tools.get_ldap_conn(
+                SERVER_OPTS['ldap_host'],
+                SERVER_OPTS['ldap_username'],
+                SERVER_OPTS['ldap_password'])
+
         result = dict()
 
         for key, value in payload.items():
             value = unquote_plus(value)
             if key == 'filter' and value == '':
-                for name, principals in all_principals:
-                    if isinstance(principals, str):
-                        result[name] = principals.split(',')
+                for name, custom_principals, realname in all_principals:
+                    if not isinstance(custom_principals, str):
+                        continue
+                    list_membership, _ = tools.get_memberof(
+                        realname,
+                        SERVER_OPTS,
+                        reuse=ldap_conn)
+                    result[name] = tools.merge_principals(
+                        custom_principals,
+                        list_membership,
+                        SERVER_OPTS).split(',')
             elif key == 'filter':
                 for principal in value.split(','):
-                    for name, principals in all_principals:
-                        if isinstance(principals, str) and principal in principals.split(','):
+                    for name, custom_principals, realname in all_principals:
+                        if not isinstance(custom_principals, str):
+                            continue
+                        list_membership, _ = tools.get_memberof(
+                            realname,
+                            SERVER_OPTS,
+                            reuse=ldap_conn)
+                        principals = tools.merge_principals(
+                            custom_principals,
+                            list_membership,
+                            SERVER_OPTS).split(',')
+                        if principal in principals:
                             if name not in result:
                                 result[name] = list()
                             result[name].append(principal)